move the !MS_IO_REDIST methods from `File.cs` to `File.netcoreapp.cs` file
